- Reorganizando arquivos de teste em subpastas por categoria dentro de /src/test - Corrige erro de rate limiting no server.js que impedia execução de múltiplos testes com npm start - No package.json adicionado novos script de testes por grupos individuais -  README.md ajustado com as novas alterações!
